refactor:[JIRATASK-000] Moved NotFound Component into common
