+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { NotFound } from '@modules/NotFound';
+import { NotFound } from '@common/NotFound';
 
 export function NotFoundPage() {
   return <NotFound />;
